# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORODIG PTS - Oro Digital Para Todos MLM Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  referralCode?: string | null;
}

export interface UpdateProfileInput {
  fullName?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  currentPassword?: string;
  /** @minLength 6 */
  newPassword?: string;
}

export type MemberRank = typeof MemberRank[keyof typeof MemberRank];


export const MemberRank = {
  Bronce: 'Bronce',
  Plata: 'Plata',
  Oro: 'Oro',
  Platino: 'Platino',
  Diamante: 'Diamante',
  Embajador: 'Embajador',
} as const;

export interface Member {
  id: number;
  username: string;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  referralCode: string;
  rank: MemberRank;
  balance: number;
  points: number;
  totalEarnings: number;
  directReferrals: number;
  totalNetwork: number;
  isActive: boolean;
  /** @nullable */
  sponsorId?: number | null;
  /** @nullable */
  sponsorName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  member: Member;
  token: string;
}

export interface EarningsByType {
  referral: number;
  sales: number;
  purchases: number;
  leadership: number;
  work: number;
  passive: number;
}

export interface MonthlyEarning {
  month: string;
  amount: number;
}

export interface DashboardSummary {
  balance: number;
  points: number;
  totalEarnings: number;
  directReferrals: number;
  totalNetwork: number;
  rank: string;
  earningsByType: EarningsByType;
  monthlyEarnings: MonthlyEarning[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  amount: number;
  createdAt: string;
}

export type EarningType = typeof EarningType[keyof typeof EarningType];


export const EarningType = {
  referral: 'referral',
  sales: 'sales',
  purchases: 'purchases',
  leadership: 'leadership',
  work: 'work',
  passive: 'passive',
} as const;

export type EarningStatus = typeof EarningStatus[keyof typeof EarningStatus];


export const EarningStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  paid: 'paid',
} as const;

export interface Earning {
  id: number;
  memberId: number;
  type: EarningType;
  description: string;
  amount: number;
  status: EarningStatus;
  /** @nullable */
  relatedMemberId?: number | null;
  /** @nullable */
  relatedMemberName?: string | null;
  createdAt: string;
}

export interface NetworkNode {
  id: number;
  fullName: string;
  username: string;
  rank: string;
  referralCode: string;
  directReferrals: number;
  totalEarnings: number;
  isActive: boolean;
  level: number;
  /** @nullable */
  avatarUrl?: string | null;
  children: NetworkNode[];
}

export interface LeaderboardEntry {
  rank: number;
  memberId: number;
  fullName: string;
  username: string;
  totalEarnings: number;
  directReferrals: number;
  memberRank: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  pointsReward: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  pointsReward: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface PurchaseInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface Purchase {
  id: number;
  memberId: number;
  productId: number;
  productName: string;
  quantity: number;
  totalPrice: number;
  pointsEarned: number;
  createdAt: string;
}

export type WithdrawalStatus = typeof WithdrawalStatus[keyof typeof WithdrawalStatus];


export const WithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  paid: 'paid',
} as const;

export interface Withdrawal {
  id: number;
  memberId: number;
  amount: number;
  method: string;
  /** @nullable */
  accountDetails?: string | null;
  status: WithdrawalStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface WithdrawalInput {
  /** @minimum 10 */
  amount: number;
  method: string;
  /** @nullable */
  accountDetails?: string | null;
}

export type AdminWithdrawalStatus = typeof AdminWithdrawalStatus[keyof typeof AdminWithdrawalStatus];


export const AdminWithdrawalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  paid: 'paid',
} as const;

export interface AdminWithdrawal {
  id: number;
  memberId: number;
  memberName: string;
  memberUsername: string;
  amount: number;
  method: string;
  /** @nullable */
  accountDetails?: string | null;
  status: AdminWithdrawalStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type AdminWithdrawalUpdateStatus = typeof AdminWithdrawalUpdateStatus[keyof typeof AdminWithdrawalUpdateStatus];


export const AdminWithdrawalUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
  paid: 'paid',
} as const;

export interface AdminWithdrawalUpdate {
  status: AdminWithdrawalUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export type AdminMemberUpdateRank = typeof AdminMemberUpdateRank[keyof typeof AdminMemberUpdateRank];


export const AdminMemberUpdateRank = {
  Bronce: 'Bronce',
  Plata: 'Plata',
  Oro: 'Oro',
  Platino: 'Platino',
  Diamante: 'Diamante',
  Embajador: 'Embajador',
} as const;

export interface AdminMemberUpdate {
  rank?: AdminMemberUpdateRank;
  isActive?: boolean;
  balance?: number;
}

export type AdminStatsRankBreakdown = {[key: string]: number};

export interface AdminStats {
  totalMembers: number;
  activeMembers: number;
  totalPaid: number;
  pendingWithdrawals: number;
  pendingAmount: number;
  totalVolume: number;
  rankBreakdown: AdminStatsRankBreakdown;
}