# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ORODIG PTS - Oro Digital Para Todos MLM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new member with optional sponsor referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error or referral code not found
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated member profile
      responses:
        "200":
          description: Current member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Unauthorized
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update current member profile (name, email, phone, password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Validation error or wrong current password
...
  /admin/withdrawals/{id}:
    patch:
      operationId: adminUpdateWithdrawal
      tags: [admin]
      summary: Admin - approve, reject or mark paid a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminWithdrawalUpdate"
      responses:
        "200":
          description: Withdrawal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin - list all products including inactive
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Admin - create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    patch:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin - update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /admin/members/{id}:
    patch:
      operationId: adminUpdateMember
      tags: [admin]
      summary: Admin - update member rank, status or balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMemberUpdate"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [username, password, fullName, email]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        referralCode:
          type: ["string", "null"]
    UpdateProfileInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      required: [member, token]
      properties:
        member:
          $ref: "#/components/schemas/Member"
        token:
          type: string
    Member:
      type: object
      required: [id, username, fullName, email, referralCode, rank, balance, points, totalEarnings, directReferrals, totalNetwork, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        referralCode:
          type: string
        rank:
          type: string
          enum: [Bronce, Plata, Oro, Platino, Diamante, Embajador]
        balance:
          type: number
        points:
          type: number
        totalEarnings:
          type: number
        directReferrals:
          type: integer
        totalNetwork:
          type: integer
        isActive:
          type: boolean
        sponsorId:
          type: ["integer", "null"]
        sponsorName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Product:
      type: object
      required: [id, name, description, price, pointsReward, category, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        pointsReward:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    ProductInput:
      type: object
      required: [name, description, price, pointsReward, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        pointsReward:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Withdrawal:
      type: object
      required: [id, memberId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        amount:
          type: number
        method:
          type: string
        accountDetails:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, paid]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    AdminWithdrawalUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, paid]
        notes:
          type: ["string", "null"]
    AdminMemberUpdate:
      type: object
      properties:
        rank:
          type: string
          enum: [Bronce, Plata, Oro, Platino, Diamante, Embajador]
        isActive:
          type: boolean
        balance:
          type: number
...