# lib/api-zod/src/generated/types/adminWithdrawalUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORODIG PTS - Oro Digital Para Todos MLM Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminWithdrawalUpdateStatus } from './adminWithdrawalUpdateStatus';

export interface AdminWithdrawalUpdate {
  status: AdminWithdrawalUpdateStatus;
  /** @nullable */
  notes?: string | null;
}