# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORODIG PTS - Oro Digital Para Todos MLM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  pointsReward: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}