# lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ORODIG PTS - Oro Digital Para Todos MLM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  fullName?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  currentPassword?: string;
  /** @minLength 6 */
  newPassword?: string;
}